Added heartbeat logic stuff to registrar
