+++ core/src/edu/rice/rbox/FaultTolerance/Registrar.java
@@ -1,21 +1,24 @@
 package edu.rice.rbox.FaultTolerance;
 
 import com.google.protobuf.Empty;
+import com.google.protobuf.Timestamp;
 import com.mongodb.client.MongoClient;
 import com.mongodb.client.MongoDatabase;
 import edu.rice.rbox.Location.Mongo.MongoManager;
-import edu.rice.rbox.Location.registrar.ClusterManager;
 import io.grpc.Server;
 import io.grpc.ServerBuilder;
 
 import io.grpc.stub.StreamObserver;
 import java.net.DatagramSocket;
 import java.net.InetAddress;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 
 
+import network.InternalRegistrarFaultToleranceGrpc;
 import network.RBoxProto;
-import network.RBoxServiceGrpc;
 import network.SuperpeerFaultToleranceGrpc.*;
 
 public class Registrar {
@@ -23,7 +26,32 @@ public class Registrar {
     private ConnectionManager connManager;
     private String ipAddress = "";
 
-    private ClusterManager clusterManager;
+    private static ClusterManager clusterManager;
+
+    private static double getNumRegistrarNodes () {
+        return (double)clusterManager.clusterMemberStubs.size();
+    }
+
+    private static boolean leader = false;
+
+    private static UUID uuid = UUID.randomUUID();
+
+    private static UUID leaderUUID;
+
+    public static UUID getUUID() {
+        return uuid;
+    }
+
+    private static RBoxProto.BasicInfo getInfo() {
+        return RBoxProto.BasicInfo.newBuilder().setSenderUUID(getUUID().toString()).setTime(getTimestamp()).build();
+    }
+
+    protected static Timestamp getTimestamp() {
+        long millis = System.currentTimeMillis();
+        Timestamp timestamp = Timestamp.newBuilder().setSeconds(millis / 1000)
+                .setNanos((int) ((millis % 1000) * 1000000)).build();
+        return timestamp;
+    }
 
     private SuperpeerFaultToleranceImplBase superPeerServiceImpl = new SuperpeerFaultToleranceImplBase() {
 
@@ -100,7 +128,7 @@ public class Registrar {
         this.connManager = new ConnectionManager(db.getCollection(MongoManager.SUPERPEER_COLLECTION),
             db.getCollection(MongoManager.CLIENT_COLLECTION));
 
-        this.clusterManager = new ClusterManager(null, new Runnable() {
+        this.clusterManager = new ClusterManager(uuid, new Runnable() {
             // this runnable notifies superpeers that this registrar is the lead
             @Override
             public void run() {
@@ -112,6 +140,9 @@ public class Registrar {
             }
         });
     }
+
+    private static Map<InternalRegistrarFaultToleranceGrpc.InternalRegistrarFaultToleranceBlockingStub, Timestamp> mostRecentClusterHeartBeats = new HashMap<>();
+    private static Map<SuperpeerFaultToleranceBlockingStub, Timestamp> mostRecentSuperpeerHeartBeats = new HashMap<>();
 
     public void init() throws  Exception {
         String ip = "";
@@ -143,6 +174,107 @@ public class Registrar {
         // Server threads are running in the background.
         System.out.println("Server started");
 
+        // create a new thread
+        Runnable target = new Runnable() {
+            @Override
+            public void run() {
+                while (!leader) {
+                    if (getTimestamp().getSeconds() > clusterManager.getMostRecentHeartbeat().getSeconds() ||
+                            (getTimestamp().getSeconds() == clusterManager.getMostRecentHeartbeat().getSeconds() &&
+                                    getTimestamp().getNanos() - clusterManager.getMostRecentHeartbeat().getNanos() > 500)) {
+                        if (clusterManager.clusterMemberStubs.size() > 1) {
+                            UUID backup = (UUID) clusterManager.clusterMemberUUIDs.keySet().toArray()[0];
+                            if (backup == leaderUUID) {
+                                backup = (UUID) clusterManager.clusterMemberUUIDs.keySet().toArray()[1];
+                            }
+                            clusterManager.clusterMemberUUIDs.get(backup).downedLeader(RBoxProto.LeaderDown.newBuilder().setSender(getInfo()).build());
+                        }
+                    }
+                }
+                while (leader) {
+                    for (InternalRegistrarFaultToleranceGrpc.InternalRegistrarFaultToleranceBlockingStub stub : clusterManager.clusterMemberStubs.keySet()) {
+                        UUID stubUUID = clusterManager.clusterMemberStubs.get(stub);
+                        RBoxProto.BasicInfo info = getInfo();
+                        RBoxProto.HeartBeatRequest req = RBoxProto.HeartBeatRequest.newBuilder().setSender(info).build();
+                        boolean success = true;
+                        try {
+                            stub.heartBeatClusterMember(req);
+                        } catch (Exception ex) {
+                            success = false;
+                        }
+                        if (success) {
+                            mostRecentClusterHeartBeats.putIfAbsent(stub, info.getTime());
+                            mostRecentClusterHeartBeats.put(stub, info.getTime());
+                        } else {
+                            if(info.getTime().getNanos() - mostRecentClusterHeartBeats.get(stub).getNanos() > 500) {
+                                int countSuccessfulChecks = 0;
+                                int countUnsuccessfulChecks = 0;
+                                for (InternalRegistrarFaultToleranceGrpc.InternalRegistrarFaultToleranceBlockingStub checkStub : clusterManager.clusterMemberStubs.keySet()) {
+                                    RBoxProto.CheckConnection check = RBoxProto.CheckConnection.newBuilder().setSender(getInfo()).setCheckUUID(stubUUID.toString()).build();
+                                    if (checkStub.checkConnectionToClusterMember(check).getResult()) {
+                                        countSuccessfulChecks++;
+                                    } else {
+                                        countUnsuccessfulChecks++;
+                                    }
+                                    if (countSuccessfulChecks > (2.0 / 3.0) * getNumRegistrarNodes()) {
+                                        break;
+                                    } else if (countUnsuccessfulChecks > (2.0 / 3.0) * getNumRegistrarNodes()) {
+                                        clusterManager.clusterMemberStubs.remove(stub);
+                                        break;
+                                    }
+                                }
+                            }
+                        }
+                    }
+                    for (SuperpeerFaultToleranceBlockingStub stub : connManager.superPeers.keySet()) {
+                        UUID stubUUID = UUID.fromString(connManager.superPeers.get(stub).getSuperPeerId());
+                        RBoxProto.BasicInfo info = getInfo();
+                        RBoxProto.HeartBeatRequest req = RBoxProto.HeartBeatRequest.newBuilder().setSender(info).build();
+                        boolean success = true;
+                        try {
+                            stub.heartBeatSuperpeer(req);
+                        } catch (Exception ex) {
+                            success = false;
+                        }
+                        if (success) {
+                            mostRecentSuperpeerHeartBeats.putIfAbsent(stub, info.getTime());
+                            mostRecentSuperpeerHeartBeats.put(stub, info.getTime());
+                        } else {
+                            if(info.getTime().getNanos() - mostRecentSuperpeerHeartBeats.get(stub).getNanos() > 500) {
+                                //TODO use mongo to get downed objects
+                                List<String> secondaryUUIDs = null;
+                                Map<String, Timestamp> secondaryBestTimestamps = null;
+                                Map<String, SuperpeerFaultToleranceBlockingStub> secondaryBestSuperpeers = null;
+                                connManager.superPeers.remove(stub);
+                                for (SuperpeerFaultToleranceBlockingStub superpeer : connManager.superPeers.keySet()) {
+                                    RBoxProto.SecondaryTimestampsMessage timestamps;
+                                    timestamps = superpeer.querySecondary(RBoxProto.QuerySecondaryMessage.newBuilder().addAllPrimaryUUIDs(secondaryUUIDs).build());
+                                    int index = 0;
+                                    for (String uuid : timestamps.getPrimaryUUIDsList()) {
+                                        secondaryBestTimestamps.putIfAbsent(uuid, timestamps.getSecondaryTimestamps(index));
+                                        secondaryBestSuperpeers.putIfAbsent(uuid, superpeer);
+                                        //TODO check this comparison
+                                        if (Integer.parseInt(secondaryBestTimestamps.get(uuid).toString()) > Integer.parseInt(timestamps.getSecondaryTimestamps(index).toString())) {
+                                            secondaryBestTimestamps.put(uuid, timestamps.getSecondaryTimestamps(index));
+                                            secondaryBestSuperpeers.put(uuid, superpeer);
+                                        }
+                                        index++;
+                                    }
+                                }
+                                //TODO double check that everything is being updated
+                                for (String uuid : secondaryBestSuperpeers.keySet()) {
+                                    RBoxProto.PromoteSecondaryMessage promote = RBoxProto.PromoteSecondaryMessage.newBuilder().setSender(getInfo()).addPromotedUUIDs(uuid).build();
+                                    secondaryBestSuperpeers.get(uuid).promote(promote);
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        };
+        Thread heartBeatChecker = new Thread(target);
+        heartBeatChecker.start();
+
         // Don't exit the main thread. Wait until server is terminated.
         server.awaitTermination();
     }
@@ -151,7 +283,6 @@ public class Registrar {
     public static void main( String[] args ) throws Exception {
         Registrar reg = new Registrar();
         reg.init();
-
 
     }
 }
